requirements.txt

Flask==3.0.2
pymongo[srv]==4.8.0
dnspython==2.6.1
python-dotenv==1.0.1
